Move task flow functions to the TaskMappingSpec
Move the task flow related functions from the WorkflowSpec to the TaskMappingSpec in preparation of delegating task related context validation to the TaskMappingSpec.
